+++ xmipp3/protocols/coordinates_tools/io_coordinates.py
@@ -1,0 +1,58 @@
+import os,sys
+from pyworkflow.utils.path import cleanPath
+from xmipp3.convert import writeSetOfCoordinates, writeMicCoordinates, readSetOfCoordinates
+from pyworkflow.em.data import SetOfCoordinates, Coordinate, Micrograph
+
+def readSetOfCoordsFromPosFnames( posDir, setOfInputCoords, sqliteOutName, write=True):
+  '''
+  posDir: path where there are .pos files with coordinates
+  setOfInputCoords. Set to find micrographs
+  setOfOutputCoordinates if not none, set where results will be written.
+  '''
+
+  inputMics = setOfInputCoords.getMicrographs()
+  micIds= inputMics.getIdSet()
+  micNameToMicId={}
+  for micId in micIds:
+    mic= inputMics[micId]
+    micNameToMicId[".".join( os.path.basename(mic.getFileName()).split(".")[:-1] )]= micId
+
+  cleanPath(sqliteOutName)
+  setOfOutputCoordinates= SetOfCoordinates(filename= sqliteOutName)
+  setOfOutputCoordinates.setMicrographs(inputMics)
+  setOfOutputCoordinates.setBoxSize( setOfInputCoords.getBoxSize())
+  readSetOfCoordinates(posDir, micSet=inputMics, coordSet=setOfOutputCoordinates, readDiscarded=False)
+  if write:
+    setOfOutputCoordinates.write()
+  return setOfOutputCoordinates
+
+
+  
+def writeCoordsListToPosFname(mic_fname, list_x_y, outputRoot):
+
+  s = """# XMIPP_STAR_1 *
+#
+data_header
+loop_
+ _pickingMicrographState
+Auto
+data_particles
+loop_
+ _xcoor
+ _ycoor
+"""
+  baseName= os.path.basename(mic_fname).split(".")[0]
+  print("%d %s %s"%(len(list_x_y), mic_fname, os.path.join(outputRoot, baseName+".pos")))
+
+  if len(list_x_y)>0:
+    with open(os.path.join(outputRoot, baseName+".pos"), "w") as f:
+        f.write(s)
+        for x, y in list_x_y:
+          f.write(" %d %d\n"%(x,y) )
+          
+def writeCoordsListToRawText(mic_fname, list_x_y, outputRoot):
+  baseName= os.path.basename(mic_fname).split(".")[0]
+  with open(os.path.join(outputRoot, baseName+"_raw_coords.txt"), "w") as f:
+      f.write("#%s %s\n"%(baseName, mic_fname))
+      for x, y in list_x_y:
+        f.write("%d %d\n"%(x, y))
