+++ xmipp3/protocols/protocol_pick_noise.py
@@ -28,18 +28,14 @@
 import os
 from pyworkflow.object import Pointer
 
-import pyworkflow.utils as pwutils
-from pyworkflow.utils.path import cleanPath
 from pyworkflow.protocol.constants import (STEPS_PARALLEL, STEPS_SERIAL, LEVEL_ADVANCED)
 import pyworkflow.protocol.params as params
 from pyworkflow.em.protocol import ProtParticlePicking
-from pyworkflow.em.data import SetOfCoordinates, Coordinate, Micrograph
+
 import pyworkflow.em.metadata as MD
-from xmipp3.convert import writeSetOfCoordinates, writeMicCoordinates, readSetOfCoordinates
+from xmipp3.convert import writeSetOfCoordinates
 from xmipp3 import XmippProtocol
-
-from pyworkflow.object import Float
-
+from xmipp3.protocols.coordinates_tools.io_coordinates import readSetOfCoordsFromPosFnames, writeCoordsListToPosFname
 
 IN_COORDS_POS_DIR_BASENAME= "pickNoiseInPosCoordinates"
 OUT_COORDS_POS_DIR_BASENAME= "pickNoiseOutPosCoordinates"
@@ -146,78 +142,6 @@ class XmippProtPickNoise(ProtParticlePicking, XmippProtocol):
         return summary
 
 
-#def readSetOfCoordsFromPosFnames( txt_coords_dirPath, setOfInputCoords, sqliteOutName=None):
-#  '''
-#  txt_coords_dirPath: path where there are txt files with coordinates
-#  setOfInputCoords. Set to find micrographs
-#  setOfOutputCoordinates if not none, set where results will be written.
-#  '''
-#  sufix="_raw_coords.txt"  
-#  inputMics = setOfInputCoords.getMicrographs()
-#  micIds= inputMics.getIdSet()
-#  micNameToMicId={}
-#  for micId in micIds:
-#    mic= inputMics[micId]
-#    micNameToMicId[".".join( os.path.basename(mic.getFileName()).split(".")[:-1] )]= micId
-
-#  if not sqliteOutName:
-#    sqliteOutName= os.path.join(txt_coords_dirPath, 'consensus_NOISE.sqlite')
-#  cleanPath('coordinates_randomPick.sqlite')
-#  cleanPath(sqliteOutName)
-
-#  setOfOutputCoordinates= SetOfCoordinates(filename= sqliteOutName)
-#  setOfOutputCoordinates.setMicrographs(inputMics)
-#  setOfOutputCoordinates.setBoxSize( setOfInputCoords.getBoxSize())
-#  nCoords=0
-#  for fname in os.listdir( txt_coords_dirPath ):
-#    if sufix is None or fname.endswith(sufix):
-#      with open(os.path.join(txt_coords_dirPath, fname)) as f:
-#        mic_name= f.readline().split()[0]
-#        try:
-#          inputMic= inputMics[micNameToMicId[mic_name]]
-#        except KeyError as e:
-#          print("Error, micName not found %s"% mic_name)
-#          continue
-#        for line in f:
-#          x_new, y_new = line.split()
-#          x_new, y_new = int(x_new), int(y_new)
-#          aux = Coordinate()
-#          aux.setMicrograph( inputMics[micNameToMicId[mic_name]] )
-#          aux.setX(x_new)
-#          aux.setY(y_new)
-#          setOfOutputCoordinates.append(aux )
-#          nCoords+=1
-#  if nCoords==0:
-#    raise Exception("No coordinates were picked. Check if the micrographs are in place") 
-#  setOfOutputCoordinates.write()
-#  return setOfOutputCoordinates
-
-
-def readSetOfCoordsFromPosFnames( posDir, setOfInputCoords, sqliteOutName, write=True):
-  '''
-  posDir: path where there are .pos files with coordinates
-  setOfInputCoords. Set to find micrographs
-  setOfOutputCoordinates if not none, set where results will be written.
-  '''
-
-  inputMics = setOfInputCoords.getMicrographs()
-  micIds= inputMics.getIdSet()
-  micNameToMicId={}
-  for micId in micIds:
-    mic= inputMics[micId]
-    micNameToMicId[".".join( os.path.basename(mic.getFileName()).split(".")[:-1] )]= micId
-
-  if write:
-    cleanPath(sqliteOutName)
-#  print("Creating output sqlite %s"%(sqliteOutName))
-  setOfOutputCoordinates= SetOfCoordinates(filename= sqliteOutName)
-  setOfOutputCoordinates.setMicrographs(inputMics)
-  setOfOutputCoordinates.setBoxSize( setOfInputCoords.getBoxSize())
-  readSetOfCoordinates(posDir, micSet=inputMics, coordSet=setOfOutputCoordinates, readDiscarded=False)
-  if write:
-    setOfOutputCoordinates.write()
-  return setOfOutputCoordinates
-
 
 def pickNoise_prepareInput(setOfCoords, outPath, outCoordsPosDir=None): 
 
@@ -238,49 +162,11 @@ def pickNoise_prepareInput(setOfCoords, outPath, outCoordsPosDir=None):
   mics_dir =  os.path.split(mic_fname)[0]
   argTuple=(boxSize, outPath, mics_dir, )
         
-
-  writeSetOfCoordinates(inCoordsPosDir, setOfCoords,
-                          scale=setOfCoords.getBoxSize())
+  writeSetOfCoordinates(inCoordsPosDir, setOfCoords)
 
   return {"boxSize":boxSize, "inCoordsPosDir": inCoordsPosDir, "outCoordsPosDir":outCoordsPosDir,
           "outPath":outPath, "mics_dir":mics_dir }
 
-#def writeCoordsListToPosFname(mic_fname, list_x_y, outputRoot, mic_id=999999):
 
-#  baseName= os.path.basename(mic_fname).split(".")[0]
-#  mic = Micrograph()
-##  mic._objId= mic_id
-#  mic.setFileName(mic_fname)
-#  coords=[]
-#  for i, (coordX, coordY) in enumerate(list_x_y):
-#    c= Coordinate()
-#    c.setX(coordX)
-#    c.setY(coordY)
-#    c.setMicrograph(mic)
-##    c._objId= i
-#    coords.append( c )
-#  print("%d %s %s"%(len(coords), mic.getFileName(), os.path.join(outputRoot, baseName+".pos")))
-#  writeMicCoordinates(mic, coords, os.path.join(outputRoot, baseName+".pos"))
-  
-def writeCoordsListToPosFname(mic_fname, list_x_y, outputRoot):
 
-  s = """# XMIPP_STAR_1 *
-#
-data_header
-loop_
- _pickingMicrographState
-Auto
-data_particles
-loop_
- _xcoor
- _ycoor
-"""
-  baseName= os.path.basename(mic_fname).split(".")[0]
-  print("%d %s %s"%(len(list_x_y), mic_fname, os.path.join(outputRoot, baseName+".pos")))
-
-  if len(list_x_y)>0:
-    with open(os.path.join(outputRoot, baseName+".pos"), "w") as f:
-        f.write(s)
-        for x, y in list_x_y:
-          f.write(" %d %d\n"%(x,y) )
